% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_features.R
\name{bind_features}
\alias{bind_features}
\title{Bind rows of feature list.}
\usage{
bind_features(feature_list)
}
\arguments{
\item{feature_list}{List of features.}
}
\value{
A data frame of features.
}
\description{
Bind rows of feature list.
}
\keyword{internal}
