% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges_to_string.R
\name{edges_to_string}
\alias{edges_to_string}
\title{Convert edges to string.}
\usage{
edges_to_string(G)
}
\arguments{
\item{G}{igraph object}
}
\value{
Character vector.
}
\description{
Convert edges to string.
}
