% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clusters.R
\name{filter_clusters}
\alias{filter_clusters}
\title{Filter a list of clusters to include only non-trivial instances.}
\usage{
filter_clusters(loc)
}
\arguments{
\item{loc}{List of clusters}
}
\value{
A subset list of clusters which contain at least one gene, one phenotype, and three members.
}
\description{
Filter a list of clusters to include only non-trivial instances.
}
