% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_edges.R
\name{get_new_edges}
\alias{get_new_edges}
\title{Get matrix of new edges.}
\usage{
get_new_edges(old_KG, new_KG)
}
\arguments{
\item{old_KG}{Old knowledge graph.}

\item{new_KG}{New knowledge graph.}
}
\value{
Character matrix.
}
\description{
Get matrix of new edges.
}
