% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features.R
\name{prepare_features}
\alias{prepare_features}
\title{Prepare the columns of a data frame based on a covariate map.}
\usage{
prepare_features(
  dt,
  feature_map,
  category,
  type = c(),
  exclusions = c(),
  verbose = TRUE
)
}
\arguments{
\item{dt}{Data frame.}

\item{feature_map}{List representing feature map with name, category, and type elements.}

\item{category}{Vector describing the order to sort the feature categories.}

\item{type}{Data types to exclude.}

\item{exclusions}{Vector of column names to exclude, regardless.}

\item{verbose}{Print final removals?}
}
\value{
Data frame
}
\description{
Prepare the columns of a data frame based on a covariate map.
}
