% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cluster.R
\name{read_cluster}
\alias{read_cluster}
\title{Reads a community files (either cluster or subcluster) as a list.}
\usage{
read_cluster(file, names = FALSE)
}
\arguments{
\item{file}{Text file where each row provides community memebers/}
}
\value{
A list where each element is a community.
}
\description{
Reads a community files (either cluster or subcluster) as a list.
}
\examples{
\dontrun{
read_cluster(file = 'data-raw/subclusters/2021/paris.greedy.2021.coms.txt', names = TRUE)
}
}
