% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clusters.R
\name{read_clusters}
\alias{read_clusters}
\title{Reads a set of community files (either clusters or subclusters) as a list.}
\usage{
read_clusters(
  path,
  pattern = "*",
  recursive = FALSE,
  names = FALSE,
  long = FALSE
)
}
\arguments{
\item{path}{String specifying where to look for all the community files.}

\item{pattern}{Regex pattern for which communities to pull.}

\item{recursive}{Logical specifying whether to recursively pull community files.}
}
\value{
A list where each element is a community.
}
\description{
Reads a set of community files (either clusters or subclusters) as a list.
}
\examples{
\dontrun{
read_clusters(
 path = 'data-raw/subclusters/2021/',
 pattern = 'paris.',
 recursive = FALSE,
 names = TRUE,
 long = FALSE
)
}
}
