% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feature.R
\name{read_feature}
\alias{read_feature}
\title{Reads a file containing cluster features as a data frame.}
\usage{
read_feature(file)
}
\arguments{
\item{file}{Text file where cluster features are written.}
}
\value{
A data frame where each row represents a cluster
}
\description{
Reads a file containing cluster features as a data frame.
}
\examples{
\dontrun{
read_feature('data-raw/features/2021/paris.greedy.2021.bocc_res.tsv')
}
}
