% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_features.R
\name{read_features}
\alias{read_features}
\title{Reads a set of files containing cluster features as a data frame.}
\usage{
read_features(path, pattern = "*", recursive = FALSE)
}
\arguments{
\item{path}{String specifying where to look for all the community files.}

\item{pattern}{Regex pattern for which communities to pull.}

\item{recursive}{Logical specifying whether to recursively pull community files.}
}
\value{
A list where each element is a community.
}
\description{
Reads a set of files containing cluster features as a data frame.
}
\examples{
read_features(
 path = 'data-raw/features/2021',
 pattern = '*',
 recursive = FALSE
)
}
