% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_KG.R
\name{read_KG}
\alias{read_KG}
\title{Reads a knowledge graph file as an igraph object.}
\usage{
read_KG(edgelist, nodes, node_attrs, node_names, cluster_args = NULL)
}
\arguments{
\item{edgelist}{Text file containing the named edgelist.}

\item{nodes}{Text file containing the node names.}

\item{node_attrs}{Text file containing the node attributes.}

\item{cluster_args}{Named list with arguments for read_clusters. Used to append the clusters as vertex attributes.}
}
\value{
A igraph object.
}
\description{
Reads a knowledge graph file as an igraph object.
}
\examples{
\dontrun{
read_KG(
 edgelist = 'data-raw/edgelists/2021/String_HPO_2021.phenotypic_branch.numbered.edgelist.txt',
 nodes = 'data-raw/edgelists/2021/String_HPO_2021.phenotypic_branch.nodenames.txt',
 node_attrs = 'data-raw/edgelists/2021/String_HPO_2021.phenotypic_branch.nodeattributes.txt'
)
}
}
