% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_network.R
\name{read_network}
\alias{read_network}
\title{Read biological network.}
\usage{
read_network(path, year)
}
\arguments{
\item{path}{Path to edgelists.}

\item{year}{Year in YYYY format.}
}
\value{
An igraph object.
}
\description{
Read biological network.
}
\details{
Reads biological network from path to edgelists and year.
}
