% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xgb_features.R
\name{read_xgb_features}
\alias{read_xgb_features}
\title{Read and format features for XGBoost}
\usage{
read_xgb_features(year)
}
\arguments{
\item{year}{Year of features in YYYY.}
}
\value{
Numeric matrix.
}
\description{
Read and format features for XGBoost
}
