% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale vector of values.}
\usage{
rescale(x, smin = 0, smax = 1)
}
\arguments{
\item{x}{Vector of positive values.}

\item{smin}{Minimum value after scaling.}

\item{smax}{Maximum value after scaling.}
}
\value{
A vector.
}
\description{
Rescale vector of values.
}
