% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ragged_tsv.R
\name{write_ragged_tsv}
\alias{write_ragged_tsv}
\title{Write ragged TSV file.}
\usage{
write_ragged_tsv(x, file)
}
\arguments{
\item{x}{A list to write to disk.}

\item{file}{File or connection to write to.}
}
\value{
A TSV file.
}
\description{
Write ragged TSV file.
}
\details{
A ragged TSV file is one with unequal number of columns per row.
}
