% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tune_from_msgs.R
\name{write_tune_from_msgs}
\alias{write_tune_from_msgs}
\title{Write tune results from msgs folder.}
\usage{
write_tune_from_msgs(array_path, msgs_path)
}
\arguments{
\item{array_path}{Folder containing the TSV files.}

\item{msgs_path}{Folder containing the msgs files.}
}
\value{
Nonempty TSV files.
}
\description{
Write tune results from msgs folder.
}
\details{
Some tune results return a segmentation fault error. When the results,
are printed to msgs, save output by reading msgs. Otherwise, keep original TSV
file.
}
